+++ src/types.ts
@@ -4,20 +4,27 @@
  * @author holyhigh
  */
 export interface RestUrl {
+  /**
+   * rest url地址，支持url参数
+   */
   url: string;
+  /**
+   * GET请求时合并到自动合并到请求参数中
+   */
+  query?: Record<string, string | number>;
   [param: string]: unknown;
 }
 
 export interface CRUDError {
-  code?: string,
-  name?: string,
-  message?: string,
-  response?: Record<string, string | number>
+  code?: string;
+  name?: string;
+  message?: string;
+  response?: Record<string, string | number>;
 }
 
 export interface Pagination {
-  pageSize: number,
-  currentPage: number,
-  total: number,
-  [propName:string]:  number
+  pageSize: number;
+  currentPage: number;
+  total: number;
+  [propName: string]: number;
 }